+++ lib/ui-drawer/ui-drawer-group.ts
@@ -30,15 +30,6 @@ class UIDrawerGroup extends LitElement {
 
     static get styles() {
         return css`
-            * {
-                box-sizing: border-box;
-            }
-
-            ::selection {
-                background-color: var(--ui-primary);
-                color: var(--ui-primary-text);
-            }
-
             :host {
                 display: block;
                 margin: 0;