+++ lib/ui-input/ui-input.ts
@@ -34,15 +34,6 @@ class UIInput extends LitElement {
 
     static get styles() {
         return css`
-            * {
-                box-sizing: border-box;
-            }
-
-            ::selection {
-                background-color: var(--ui-primary);
-                color: var(--ui-primary-text);
-            }
-
             :host {
                 display: block;