+++ lib/ui-input/ui-search.ts
@@ -39,15 +39,6 @@ class UISearch extends LitElement {
 
     static get styles() {
         return css`
-            * {
-                box-sizing: border-box;
-            }
-
-            ::selection {
-                background-color: var(--ui-primary));
-                color: var(--ui-primary-text);
-            }
-
             :host {
                 display: block;