+++ lib/ui-input/ui-textarea.ts
@@ -32,15 +32,6 @@ class UITextarea extends LitElement {
 
     static get styles() {
         return css`
-            * {
-                box-sizing: border-box;
-            }
-
-            ::selection {
-                background-color: var(--ui-primary);
-                color: var(--ui-primary-text);
-            }
-
             :host {
                 display: block;
                 position: relative;