+++ lib/ui-text/ui-label.ts
@@ -1,6 +1,5 @@
-import { css, html, LitElement, PropertyValues } from "lit";
+import { css, html, LitElement } from "lit";
 import { customElement, property } from "lit/decorators.js";
-import { CleanUpFunction, ripple } from "..";
 
 /**
  * @element ui-label
@@ -9,44 +8,14 @@ import { CleanUpFunction, ripple } from "..";
  */
 @customElement("ui-label")
 class UILabel extends LitElement {
-    private rippleCleanUp: CleanUpFunction | null = null;
-
-    private clickHandler = async (ev: Event) => {
-        if (!ev.currentTarget) return;
-
-        const container = ev.currentTarget as HTMLElement;
-        ([...container.children] as HTMLElement[]).forEach((child) => {
-            child.click();
-        });
-    };
-
-    private childClickHandler = async (ev: Event) => {
-        ev.stopPropagation();
-    };
-
     @property({ type: String, attribute: "primary", reflect: true })
     primary?: string;
 
     @property({ type: String, attribute: "secondary", reflect: true })
     secondary?: string;
 
-    /**
-     * All slotted elements will be clicked once if ripple was set to true
-     */
-    @property({ type: Boolean, attribute: "ripple", reflect: true })
-    ripple: boolean = false;
-
     static get styles() {
         return css`
-            * {
-                box-sizing: border-box;
-            }
-
-            ::selection {
-                background-color: var(--ui-primary);
-                color: var(--ui-primary-text);
-            }
-
             :host {
                 display: block;
             }
@@ -94,61 +63,6 @@ class UILabel extends LitElement {
                 </span>
             </div>
         `;
-    }
-
-    protected firstUpdated(_changedProperties: PropertyValues): void {
-        super.firstUpdated(_changedProperties);
-
-        if (this.ripple) this.enableRipple();
-        else this.disableRipple();
-    }
-
-    attributeChangedCallback(name: string, _old: string | null, value: string | null): void {
-        super.attributeChangedCallback(name, _old, value);
-
-        switch (name) {
-            case "ripple":
-                if (this.ripple) {
-                    this.enableRipple();
-                } else {
-                    this.disableRipple();
-                }
-        }
-    }
-
-    private async enableRipple() {
-        if (!!this.rippleCleanUp) return;
-
-        const container = this.shadowRoot?.querySelector<HTMLElement>(`div`);
-        if (!container) {
-            // Component isn't rendered yet!
-            return;
-        }
-
-        this.rippleCleanUp = ripple.create(container);
-
-        this.role = "button";
-        this.style.cursor = "pointer";
-
-        this.addEventListener("click", this.clickHandler);
-        ([...this.children] as HTMLElement[]).forEach((child) => {
-            child.addEventListener("click", this.childClickHandler);
-        });
-    }
-
-    private async disableRipple() {
-        if (!this.rippleCleanUp) return;
-
-        this.style.cursor = "default";
-        this.role = null;
-
-        this.rippleCleanUp();
-        this.rippleCleanUp = null;
-
-        this.removeEventListener("click", this.clickHandler);
-        ([...this.children] as HTMLElement[]).forEach((child) => {
-            child.removeEventListener("click", this.childClickHandler);
-        });
     }
 }